requirements.txt
llama-cpp-python
spacy
nltk
spacy-entity-linker
wikipedia-api